Move CTA slice to footer only.
