+++ src/layout/footer/style.css
@@ -1,5 +1,5 @@
-.pre-footer-slice {
-  border-top: 3px solid var(--pds-color-background-default);
+.pds-site-footer {
+  border-top: 1px solid var(--pds-color-border-default);
 }
 
 .footer-form-intro {
