+++ src/pages/index.js
@@ -9,7 +9,18 @@ import LinksList from "../pds-middleware/LinksList"
 import CallToAction from "../pds-middleware/CallToAction"
 import TopicsGrid from "../pds-middleware/TopicsGrid"
 
-import { TwoItemLayout } from "@pantheon-systems/pds-toolkit-react"
+import { CTASlice, TwoItemLayout } from "@pantheon-systems/pds-toolkit-react"
+
+// Content for CTA
+const primaryCTA = {
+  text: "Learn Pantheon",
+  url: "https://pantheon.io/learn-pantheon?docs",
+}
+
+const secondaryCTA = {
+  text: "Office Hours",
+  url: "https://pantheon.io/developers/office-hours?docs",
+}
 
 class Index extends React.Component {
   render() {
@@ -66,6 +77,14 @@ class Index extends React.Component {
             />
           </Helmet>
         </main>
+        <CTASlice
+          backgroundColor="secondary"
+          containerWidth="wide"
+          headingText="Got questions? We've got answers!"
+          primaryLink={primaryCTA}
+          secondaryLink={secondaryCTA}
+          className="pre-footer-slice"
+        />
       </Layout>
     )
   }
